test/com/homedepot/di/xd/efs/dao/ReportTypeDAOTest.java
package com.homedepot.di.xd.efs.dao;

import java.util.List;

import org.junit.Test;

import org.junit.Assert;

import com.homedepot.di.xd.efs.exception.DataOperationException;
import com.homedepot.di.xd.efs.to.ReportTypeTO;
import com.homedepot.homer.util.HomerUnitTestCase;

public class ReportTypeDAOTest extends HomerUnitTestCase{
	
	//Test that we can retrieve all report type codes successfully.
	@Test
	public void testRetrieveAllReportTypeCodes_Success() throws DataOperationException{
		List<ReportTypeTO> reportTypesList = ReportTypeDAO.getReportTypeCodes();
		
		Assert.assertNotNull(reportTypesList);
		Assert.assertEquals(2, reportTypesList.size());

		 Assert.assertTrue(true);
	}

}
